4. Container orchestration and serverless/http-function/requirements.txt
Flask==2.3.2
google-cloud-storage==2.10.0
functions-framework
